requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
qdrant-client==1.13.3
openai==1.82.0
httpx==0.28.1
pydantic==2.11.1
